requirements.txt
certifi==2024.8.30
influxdb-client==1.46.0
PyQt6==6.7.1
PyQt6-Qt6==6.7.3
PyQt6_sip==13.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
reactivex==4.0.4
setuptools==75.2.0
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.3
